refactor loginservice to folder
